cambio de variable

+++ src/main/java/bo/ucb/edu/ingsoft/dto/VoucherAdd.java
@@ -49,7 +49,7 @@ public class VoucherAdd {
         this.date = date;
     }
 
-    public Float getTotal() {
+    public BigDecimal getTotal() {
         return total;
     }
